+++ packages/sdk/src/Platform/MetaMaskInstaller.ts
@@ -1,7 +1,7 @@
 import MetaMaskOnboarding from '@metamask/onboarding';
+import { PlatformType } from '@metamask/sdk-communication-layer';
 import { Ethereum } from '../services/Ethereum';
 import { ProviderService } from '../services/ProviderService';
-import { PlatformType } from '../types/PlatformType';
 import { waitPromise } from '../utils/waitPromise';
 import { Platform } from './Platfform';
 
@@ -58,7 +58,7 @@ export class MetaMaskInstaller {
   }
 
   startDesktopOnboarding() {
-    if (this.debug) {
+    if (this?.debug) {
       console.debug(`MetamaskInstaller::startDesktopOnboarding()`);
     }
     Ethereum.destroy();
@@ -70,7 +70,7 @@ export class MetaMaskInstaller {
   async redirectToProperInstall() {
     const platformType = Platform.getInstance().getPlatformType();
 
-    if (this.debug) {
+    if (this?.debug) {
       console.debug(
         `MetamaskInstaller::redirectToProperInstall() platform=${platformType} this.preferDesktop=${this.preferDesktop}`,
       );