+++ packages/sdk/src/services/WalletConnect.ts
@@ -1,6 +1,5 @@
-import { ChannelConfig } from '@metamask/sdk-communication-layer';
+import { ChannelConfig, PlatformType } from '@metamask/sdk-communication-layer';
 import { Platform } from '../Platform/Platfform';
-import { PlatformType } from '../types/PlatformType';
 import InstallModal from '../ui/InstallModal/installModal';
 import { ProviderService } from './ProviderService';
 
@@ -105,6 +104,10 @@ export class WalletConnect implements ProviderService {
 
   isConnected() {
     return this.wcConnector.connected;
+  }
+
+  isAuthorized(): boolean {
+    throw new Error('Method not implemented.');
   }
 
   disconnect(): void {